reorganized to Maven project structure
