--- Java/com/gildedrose/Item.java
@@ -1,21 +1,0 @@
-package com.gildedrose;
-
-public class Item {
-
-    public String name;
-
-    public int sellIn;
-
-    public int quality;
-
-    public Item(String name, int sellIn, int quality) {
-        this.name = name;
-        this.sellIn = sellIn;
-        this.quality = quality;
-    }
-
-   @Override
-   public String toString() {
-        return this.name + ", " + this.sellIn + ", " + this.quality;
-    }
-}
