+++ Java/src/test/java/com.gildedrose/GildedRoseTest.java
@@ -1,0 +1,17 @@
+package com.gildedrose;
+
+import static org.junit.Assert.*;
+
+import org.junit.Test;
+
+public class GildedRoseTest {
+
+    @Test
+    public void foo() {
+        Item[] items = new Item[] { new Item("foo", 0, 0) };
+        GildedRose app = new GildedRose(items);
+        app.updateQuality();
+        assertEquals("fixme", app.items[0].name);
+    }
+
+}
